# Conflicts:
#	mag_jang/build.gradle

+++ mag_jang/src/main/java/B208/mag_jang/service/UserService.java
@@ -17,7 +17,7 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public Integer join(User user){
+    public Long join(User user){
         validateDuplicateUser(user);
 
         userRepository.save(user);
